% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_clustering}
\alias{check_clustering}
\title{Check the validity of clustering.}
\usage{
check_clustering(clustering, size_constraint)
}
\arguments{
\item{clustering}{a \code{Rscc_clustering} object containing an existing non-empty clustering.}

\item{size_constraint}{an integer with the required minimum cluster size.}
}
\value{
Returns \code{TRUE} if \code{clustering} satisfies the size constraint, and \code{FALSE}
        if it does not. Throws an error if \code{clustering} is an invalid clustering.
}
\description{
\code{check_clustering} checks so the inputted clustering satisfies the specified size
constraint.
}
\examples{
# Each cluster contains at least two data
# points; it's valid for `size_constraint == 2`.
my_clust_obj1 <- Rscc_clustering(c("A", "A", "B", "C",
                                   "B", "C", "C", "A",
                                   "B", "B"))
check_clustering(my_clust_obj1, 2)
# > TRUE


# One cluster contains only one point. This is
# an invalid clustering for `size_constraint == 2`.
my_clust_obj2 <- Rscc_clustering(c("A", "A", "B", "C",
                                   "B", "C", "C", "A",
                                   "B", "B", "D"))
check_clustering(my_clust_obj2, 2)
# > FALSE


# `my_clust_obj1` doesn't contain clusters with
# at least 5 units, it's invalid for `size_constraint == 2`.
check_clustering(my_clust_obj1, 5)
# > FALSE

}
\seealso{
\code{\link{check_clustering_types}} checks type-specific size constraints in addition
         to an overall size constraint.
}

