% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_scclust}
\alias{check_scclust}
\title{Check clustering constraints}
\usage{
check_scclust(clustering, size_constraint = NULL, type_labels = NULL,
  type_constraints = NULL)
}
\arguments{
\item{clustering}{a \code{\link{scclust}} object containing a non-empty clustering.}

\item{size_constraint}{an integer with the required minimum cluster size. If \code{NULL},
only the type constraints will be checked.}

\item{type_labels}{a vector containing the type of each data point. May be
\code{NULL} when \code{type_constraints} is \code{NULL}.}

\item{type_constraints}{a named integer vector containing type-specific size constraints.
If \code{NULL}, only the overall constraint will be checked.}
}
\value{
Returns \code{TRUE} if \code{clustering} satisfies the clustering constraints, and
        \code{FALSE} if it does not. Throws an error if \code{clustering} is an invalid
        instance of the \code{\link{scclust}} class.
}
\description{
\code{check_scclust} checks whether the clusters in a clustering satisfies a set
of conditions on their size and composition.
}
\examples{
# Example scclust clustering
my_scclust <- scclust(c("A", "A", "B", "C", "B",
                        "C", "C", "A", "B", "B"))


# Check so each cluster contains at least two data points
check_scclust(my_scclust, 2)
# > TRUE


# Check so each cluster contains at least three data points
check_scclust(my_scclust, 3)
# > FALSE


# Data point types
my_types <- factor(c("x", "y", "y", "z", "z",
                     "x", "y", "z", "x", "x"))


# Check so each cluster contains at least one point of each type
check_scclust(my_scclust,
              NULL,
              my_types,
              c("x" = 1, "y" = 1, "z" = 1))
# > TRUE


# Check so each cluster contains one point of each type
# and at least three points in total
check_scclust(my_scclust,
              3,
              my_types,
              c("x" = 1, "y" = 1, "z" = 1))
# > TRUE


# Check so each cluster contains five data points of type "y"
check_scclust(my_scclust,
              NULL,
              my_types,
              c("y" = 5))
# > FALSE


# Check so each cluster contains one data point of both "x" and "z"
# and at least three points in total
check_scclust(my_scclust,
              3,
              my_types,
              c("x" = 1, "z" = 1))
# > TRUE

}
\seealso{
See \code{\link{sc_clustering}} for details on how to specify the
         \code{type_labels} and \code{type_constraints} parameters.
}
