% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scclust.R
\name{cluster_count}
\alias{cluster_count}
\title{Count the number of clusters in a clustering}
\usage{
cluster_count(clustering)
}
\arguments{
\item{clustering}{a \code{\link{scclust}} object containing a non-empty clustering.}
}
\value{
Returns an integer with the number of clusters in \code{clustering}.
}
\description{
\code{cluster_count} returns the number of clusters in a clustering.
}
\examples{
# Example scclust clustering
my_scclust <- scclust(c("A", "A", "B", "C", "B",
                        "C", "C", "A", "B", "B"))

cluster_count(my_scclust)
# > 3

}
