% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_point_count.R
\name{data_point_count}
\alias{data_point_count}
\title{Number of data points}
\usage{
data_point_count(x)
}
\arguments{
\item{x}{an Rscc_clustering or Rscc_distances object}
}
\value{
Returns an integer with the number of data points in \code{x}.
}
\description{
\code{data_point_count} returns the number of data points in a clustering
or distance object.
}

