% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{get_scclust_stats}
\alias{get_scclust_stats}
\title{Derive statistics about a clustering.}
\usage{
get_scclust_stats(clustering, distances)
}
\arguments{
\item{clustering}{a \code{\link{scclust}} object containing a non-empty clustering.}

\item{distances}{a \code{\link[distances]{distances}} object describing the distances
between the data points in \code{clustering}.}
}
\value{
Returns a list of class \code{scclust_stats} containing the statistics.
}
\description{
\code{get_scclust_stats} calculates statistics about a clustering.
}
\details{
The function reports the following statistics:

\tabular{ll}{
  \code{num_data_points} \tab The total number of data points. \cr
  \code{num_assigned} \tab The number of points assigned to a cluster. \cr
  \code{num_clusters} \tab The number of clusters. \cr
  \code{min_cluster_size} \tab The size of the smallest cluster. \cr
  \code{max_cluster_size} \tab The size of the largest cluster. \cr
  \code{avg_cluster_size} \tab The average cluster size. \cr
  \code{sum_dists} \tab The sum of all within-cluster distances. \cr
  \code{min_dist} \tab The overall smallest within-cluster distance. \cr
  \code{max_dist} \tab The overall largest within-cluster distance. \cr
  \code{cl_avg_min_dist} \tab The average of the clusters' smallest distance. \cr
  \code{cl_avg_max_dist} \tab The average of the clusters' largest distance. \cr
  \code{cl_avg_dist_weighted} \tab The average of the clusters' average distance weighed by cluster size. \cr
  \code{cl_avg_dist_unweighted} \tab The average of the clusters' average distance (unweighed). \cr
}

Let \eqn{d(i,j)}{d(i,j)} denote the distance between data points \eqn{i}{i} and \eqn{j}{j}. Let \eqn{c}{c} be a cluster
containing all indices of points assigned to the cluster. Let
\deqn{D(c) = \{d(i,j): i,j \in c \wedge i>j\}}{D(c) = { d(i,j) : i,j in c and i > j }}
be a set function returning all unique within-cluster distances in \eqn{c}{c}. Let \eqn{C}{C} be a set containing
all clusters.

\code{sum_dists} is defined as:
\deqn{\sum_{c\in C} sum(D(c))}{\sum_[c in C] sum(D(c))}

\code{min_dist} is defined as:
\deqn{\min_{c\in C} \min(D(c))}{min_[c in C] min(D(c))}

\code{max_dist} is defined as:
\deqn{\max_{c\in C} \max(D(c))}{max_[c in C] max(D(c))}

\code{cl_avg_min_dist} is defined as:
\deqn{\sum_{c\in C} \frac{\min(D(c))}{|C|}}{\sum_[c in C] min(D(c)) / count(C)}

\code{cl_avg_max_dist} is defined as:
\deqn{\sum_{c\in C} \frac{\max(D(c))}{|C|}}{\sum_[c in C] max(D(c)) / count(C)}

Let:
\deqn{AD(c) = \frac{sum(D(c))}{|D(c)|}}{AD(c) = sum(D(c)) / count(D(c))}
be the average within-cluster distance in cluster \eqn{c}{c}.

\code{cl_avg_dist_weighted} is defined as:
\deqn{\sum_{c\in C} \frac{|c| AD(c)}{num_assigned}}{\sum_[c in C] count(c) * AD(c) / num_assigned}
where \eqn{num_assigned}{num_assigned} is the number of assigned data points (see above).

\code{cl_avg_dist_unweighted} is defined as:
\deqn{\sum_{c\in C} \frac{AD(c)}{|C|}}{\sum_[c in C] AD(c) / count(C)}
}
\examples{
my_scclust <- scclust(c("A", "A", "B", "C", "B",
                        "C", "C", "A", "B", "B"))

my_data_points <- data.frame(x = c(0.1, 0.2, 0.3, 0.4, 0.5,
                                   0.6, 0.7, 0.8, 0.9, 1.0),
                             y = c(10, 9, 8, 7, 6,
                                   10, 9, 8, 7, 6))

my_distances <- distances(my_data_points)

get_scclust_stats(my_scclust, my_distances)

# >                        Value
# > num_data_points        10.0000000
# > num_assigned           10.0000000
# > num_clusters            3.0000000
# > min_cluster_size        3.0000000
# > max_cluster_size        4.0000000
# > avg_cluster_size        3.3333333
# > sum_dists              18.2013097
# > min_dist                0.5000000
# > max_dist                3.0066593
# > cl_avg_min_dist         0.8366584
# > cl_avg_max_dist         2.4148611
# > cl_avg_dist_weighted    1.5575594
# > cl_avg_dist_unweighted  1.5847484

}
