% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_clustering.R
\name{hierarchical_clustering}
\alias{hierarchical_clustering}
\title{Hierarchical clustering function}
\usage{
hierarchical_clustering(distance_object, size_constraint, batch_assign = TRUE,
  existing_clustering = NULL)
}
\arguments{
\item{distance_object}{a distance object as produced by \code{\link{make_distances}}.}

\item{size_constraint}{an integer with the required minimum cluster size.}

\item{batch_assign}{a bool indicating whether data points should be assigned in batches when
spliting clusters.}

\item{existing_clustering}{\code{NULL} or a \code{Rscc_clustering} object containing an existing
non-empty clustering.}
}
\value{
Returns a Rscclust cluster object containing the derived clustering.
}
\description{
\code{hierarchical_clustering} derives a clustering statisfying a specified
size-constraint using a hierarchical clustering algorithm. The primary purpose of
this function is to refine clusterings produced by the nearest neighbor graph
algorithms in the packages.
}
\details{
To be written...
}
\seealso{
\code{\link{nng_clustering}}, \code{\link{nng_clustering_batches}} and
         \code{\link{nng_clustering_types}} are other clustering functions in the package.

         Use \code{\link{Rscc_clustering}} to create Rscclust cluster objects from external
         clusterings.

Other clustering.functions: \code{\link{nng_clustering}}
}
\keyword{cluster}

