% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_clustering.R
\name{hierarchical_clustering_internal}
\alias{hierarchical_clustering_internal}
\title{Internal hierarchical clustering function}
\usage{
hierarchical_clustering_internal(distance_object, size_constraint,
  batch_assign = TRUE, existing_clustering = NULL, deep_copy = TRUE)
}
\arguments{
\item{distance_object}{a distance object as produced by \code{\link{make_distances}}.}

\item{size_constraint}{an integer with the required minimum cluster size.}

\item{batch_assign}{a bool indicating whether data points should be assigned in batches when
spliting clusters.}

\item{existing_clustering}{\code{NULL} or a \code{Rscc_clustering} object containing an existing
non-empty clustering. If \code{NULL}, the function will start with a
single cluster containing all data points (i.e., it derives a clustering
from scratch).}

\item{deep_copy}{a bool indicating whether a deep copy of
\code{existing_clustering} should be made before
refining the clustering.}
}
\value{
Returns a Rscclust cluster object containing the derived clustering.
}
\description{
ATTENTION! Using this function has side-effects on your R environment,
ensure you properly understand what this does before using it.
In most cases, \code{\link{hierarchical_clustering}} is the preferred
choice.
}
\details{
\code{hierarchical_clustering_internal} is identical to
\code{\link{hierarchical_clustering}} except for the
\code{deep_copy} parameter. When \code{deep_copy == TRUE}, the
two functions are identical. However, when
\code{deep_copy == FALSE}, \code{hierarchical_clustering_internal}
will \strong{not} make a deep copy of \code{existing_clustering} before
refining the clustering. That is, any changes will be made in place, and
\code{existing_clustering} will be invalidated by calling
this function. This is useful when refining an
existing clustering and the inputted clustering is not of interest.

See \code{\link{hierarchical_clustering}} for detailed documentation.
}
\keyword{internal}

