% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rscc_clustering.R
\name{is.Rscc_clustering}
\alias{is.Rscc_clustering}
\title{Check \code{Rscc_clustering} object}
\usage{
is.Rscc_clustering(obj)
}
\arguments{
\item{obj}{object to check.}
}
\value{
Returns \code{TRUE} if \code{obj} is a valid
        \code{Rscc_clustering} object, otherwise \code{FALSE}.
}
\description{
\code{is.Rscc_clustering} checks whether the provided object
is a valid instance of the \code{Rscc_clustering} class. It does
not whether the clustering itself is sensible.
}

