% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scc_clustering.R
\name{is.scc_clustering}
\alias{is.scc_clustering}
\title{Check \code{scc_clustering} object}
\usage{
is.scc_clustering(obj)
}
\arguments{
\item{obj}{object to check.}
}
\value{
Returns \code{TRUE} if \code{obj} is a valid
        \code{scc_clustering} object, otherwise \code{FALSE}.
}
\description{
\code{is.scc_clustering} checks whether the provided object
is a valid instance of the \code{scc_clustering} class. It does
not check whether the clustering itself is sensible.
}
