% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distances.R
\name{is.scc_distances}
\alias{is.scc_distances}
\title{Check \code{scc_distances} object}
\usage{
is.scc_distances(obj)
}
\arguments{
\item{obj}{object to check.}
}
\value{
Returns \code{TRUE} if \code{obj} is a valid
        \code{scc_distances} object, otherwise \code{FALSE}.
}
\description{
\code{is.scc_distances} checks whether the provided object
is a valid instance of the \code{scc_distances} class.
}
