% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nng_clustering.R
\name{nng_clustering}
\alias{nng_clustering}
\title{Nearest neighbor graph based clustering function}
\usage{
nng_clustering(distance_object, size_constraint,
  seed_method = "exclusion_updating",
  main_unassigned_method = "closest_seed", main_radius = NULL,
  main_data_points = NULL, secondary_unassigned_method = "ignore",
  secondary_radius = NULL)
}
\arguments{
\item{distance_object}{a distance object as produced by \code{\link{make_distances}}.}

\item{size_constraint}{an integer with the required minimum cluster size.}

\item{seed_method}{how seeds are found. See below for additional details.}

\item{main_unassigned_method}{how remaining (primary) data points are assigned
to clusters. See below for additional details.}

\item{main_radius}{restricts the maximum length of an edge in the graph that
summaries the distance information. \code{NULL} indicates
no restriction. See below for details on how this translates
to maximum distance within a cluster.}

\item{main_data_points}{a logical vector of length equal to number of data points
indicating primary and secondary data points. \code{NULL}
indicates that all data points are "primary". See below for
details.}

\item{secondary_unassigned_method}{how remaining secondary data points are assigned
to clusters. See below for additional details.}

\item{secondary_radius}{restricts the maximum distance when assigning secondary data
points to clusters.}
}
\value{
Returns a Rscclust cluster object containing the derived clustering.
}
\description{
\code{nng_clustering} derives a clustering statisfying the specified
size constraints. It implements an algorithm that first summaries the
distance information between data points in a sparse graph and then
constructs the clusterings based on the graph. The function is intended
to run fast while ensuring near-optimal performance.
}
\details{
Must be one of
                   "lexical", "inwards_order", "inwards_updating",
                   "inwards_alt_updating", "exclusion_order", "exclusion_updating".

Must be one of
                              "ignore", "by_nng", "closest_assigned",
                              "closest_seed", "estimated_radius_closest_seed".
}
\seealso{
\code{\link{nng_clustering_batches}} and \code{\link{nng_clustering_types}} are
         other NNG-based clustering functions in the package.

         \code{\link{hierarchical_clustering}} can be used to refine the clustering
         constructed by this function.

Other clustering.functions: \code{\link{hierarchical_clustering}}
}
\keyword{cluster}

