% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nng_clustering.R
\name{nng_clustering_types}
\alias{nng_clustering_types}
\title{Dummy documentation}
\usage{
nng_clustering_types(distance_object, type_labels, type_size_constraints,
  total_size_constraint = NULL, seed_method = "exclusion_updating",
  main_unassigned_method = "closest_seed", main_radius = NULL,
  main_data_points = NULL, secondary_unassigned_method = "ignore",
  secondary_radius = NULL)
}
\arguments{
\item{distance_object}{abc ...}

\item{type_labels}{abc ...}

\item{type_size_constraints}{abc ...}

\item{total_size_constraint}{abc ...}

\item{seed_method}{abc ...}

\item{main_unassigned_method}{abc ...}

\item{main_radius}{abc ...}

\item{main_data_points}{abc ...}

\item{secondary_unassigned_method}{abc ...}

\item{secondary_radius}{abc ...}
}
\description{
Dummy documentation
}

