% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rscc_clustering.R
\name{Rscc_clustering}
\alias{Rscc_clustering}
\title{Constructor for Rscclust clustering objects.}
\usage{
Rscc_clustering(cluster_labels, unassigned_labels = NULL, ids = NULL)
}
\arguments{
\item{cluster_labels}{a vector containing each data point's cluster label.}

\item{unassigned_labels}{a vector containing cluster labels that denote unassigned data points.}

\item{ids}{optional IDs of the data points. Should be a vector of the same length as
\code{cluster_labels} or \code{NULL}. If \code{NULL}, the IDs are set to
\code{1:length(cluster_labels)}.}
}
\value{
Returns a Rscclust cluster object to be used when calling the
        clustering functions.
}
\description{
\code{Rscc_clustering} constructs a Rscclust clustering object from existing cluster
labels. It is useful when using Rscclust's clustering functions to refine an
existing clustering.
}
\examples{
# 10 data points in 3 clusters
my_labels1 <- c("A", "A", "B", "C", "B", "C", "C", "A", "B", "B")
my_clust_obj1 <- Rscc_clustering(my_labels1)

# 10 data points in 3 clusters, integer labels
my_labels2 <- c(1, 1, 2, 3, 2, 3, 3, 1, 2, 2)
my_clust_obj2 <- Rscc_clustering(my_labels2)

# 8 data points in 3 clusters, and 2 points unassigned
my_labels3 <- c("A", "A", "B", "C", "NONE", "C", "C", "NONE", "B", "B")
my_clust_obj3 <- Rscc_clustering(my_labels3, "NONE")

# Different unassiged labels
my_labels4 <- c("A", "A", "B", "C", "NONE", "C", "C", "0", "B", "B")
my_clust_obj4 <- Rscc_clustering(my_labels4, c("NONE", "0"))

# Custom data point IDs
my_labels5 <- c("A", "A", "B", "C", "B", "C", "C", "A", "B", "B")
my_ids <- letters[1:10]
my_clust_obj5 <- Rscc_clustering(my_labels5, ids = my_ids)

}
\seealso{
Functions that accept the Rscclust cluster object as input are
\code{\link{hierarchical_clustering}}, \code{\link{check_clustering}},
\code{\link{check_clustering_types}} and \code{\link{get_clustering_stats}} for

\code{Rscc_clustering} does not derive clusterings from data sets; for that
functionality see \code{\link{make_clustering}} and \code{\link{hierarchical_clustering}}.
}

