% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rscclust.R
\docType{package}
\name{Rscclust-package}
\alias{Rscclust}
\alias{Rscclust-package}
\title{Rscclust: R wrapper for the scclust library}
\description{
This package wraps the \code{scclust} C library for R. \code{scclust}
provides functions for size constrained clustering.
}
\details{
This package is under heavy development, please exercise great caution when using it.

More information and the latest version is found here:
\url{https://github.com/fsavje/Rscclust}.

More information about the underlying \code{scclust} library is found here:
\url{https://github.com/fsavje/scclust}.

Bug reports and suggestions are greatly appreciated. They
are best reported here:
\url{https://github.com/fsavje/Rscclust/issues}.
}

