% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{set_dist_functions}
\alias{set_dist_functions}
\title{Set distance search functions.}
\usage{
set_dist_functions(dist_functions = "ann")
}
\arguments{
\item{dist_functions}{a string containing either "ann" or "internal".}
}
\value{
Returns \code{NULL}.
}
\description{
\code{set_dist_functions} sets the distance search functions used by scclust to construct
clusterings. The package is loaded with the distance functions set to ANN, i.e., equivalent
to running \code{set_dist_functions()} when the package loads.
}
\examples{
set_dist_functions()
set_dist_functions(dist_functions = "ann")
set_dist_functions(dist_functions = "internal")

}
