% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rscc_clustering.R
\name{cluster_count}
\alias{cluster_count}
\title{Number of clusters in clustering}
\usage{
cluster_count(clustering)
}
\arguments{
\item{clustering}{an Rscc_clustering object}
}
\value{
Returns an integer with the number of clusters in \code{clustering}.
}
\description{
\code{cluster_count} returns the number of clusters in a clustering object.
}

