% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scclust.R
\name{is.scclust}
\alias{is.scclust}
\title{Check \code{scclust} object}
\usage{
is.scclust(obj)
}
\arguments{
\item{obj}{object to check.}
}
\value{
Returns \code{TRUE} if \code{obj} is a valid
        \code{scclust} object, otherwise \code{FALSE}.
}
\description{
\code{is.scclust} checks whether the provided object
is a valid instance of the \code{scclust} class. It does
not check whether the clustering itself is sensible.
}
