% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nng_clustering.R
\name{nng_clustering}
\alias{nng_clustering}
\title{Dummy documentation}
\usage{
nng_clustering(distance_object, size_constraint,
  seed_method = "exclusion_updating",
  main_unassigned_method = "closest_seed", main_radius = NULL,
  main_data_points = NULL, secondary_unassigned_method = "ignore",
  secondary_radius = NULL)
}
\arguments{
\item{distance_object}{abc ...}

\item{size_constraint}{abc ...}

\item{seed_method}{abc ...}

\item{main_unassigned_method}{abc ...}

\item{main_radius}{abc ...}

\item{main_data_points}{abc ...}

\item{secondary_unassigned_method}{abc ...}

\item{secondary_radius}{abc ...}
}
\description{
Dummy documentation
}
\seealso{
Other clustering.functions: \code{\link{top_down_greedy_clustering}}
}
\keyword{cluster}

