% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rscclust.R
\docType{package}
\name{Rscclust-package}
\alias{Rscclust}
\alias{Rscclust-package}
\title{Rscclust: R wrapper for the scclust library}
\description{
This package wraps the \code{scclust} C library for R. \code{scclust}
provides functions for size constrained clustering. Besides direct wrapper
function, the package provides convenience and utility functions.
}
\details{
The clustering functions are \code{\link{make_clustering}} and
\code{\link{hierarchical_clustering}}. To get statistics about a clustering
use \code{\link{get_clustering_stats}}. To construct a clustering object from
an existing clustering see \code{\link{Rscc_clustering}}. To check if a clustering
satisfies some clustering constraints use \code{\link{check_clustering}}.

This package is under development, please exercise caution when using it.

More information and the latest version is found here:
\url{https://github.com/fsavje/Rscclust}.

More information about the underlying \code{scclust} library is found here:
\url{https://github.com/fsavje/scclust}.

Bug reports and suggestions are greatly appreciated. They
are best reported here:
\url{https://github.com/fsavje/Rscclust/issues}.
}

