% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_clustering.R
\name{sc_clustering}
\alias{sc_clustering}
\title{Size constrained clustering}
\usage{
sc_clustering(distances, size_constraint = NULL, type_labels = NULL,
  type_constraints = NULL, seed_method = "exclusion_updating",
  primary_data_points = NULL, primary_unassigned_method = "closest_seed",
  secondary_unassigned_method = "ignore", seed_radius = NULL,
  primary_radius = "seed_radius", secondary_radius = NULL,
  batch_size = 100L)
}
\arguments{
\item{distances}{a distance object as produced by \code{\link[distances]{distances}}.}

\item{size_constraint}{an integer with the required minimum cluster size.}

\item{type_labels}{...}

\item{type_constraints}{...}

\item{seed_method}{how seeds are found. See below for additional details.}

\item{primary_data_points}{a logical vector of length equal to number of data points
indicating primary and secondary data points. \code{NULL}
indicates that all data points are "primary". See below for
details.}

\item{primary_unassigned_method}{how remaining (primary) data points are assigned
to clusters. See below for additional details.}

\item{secondary_unassigned_method}{how remaining secondary data points are assigned
to clusters. See below for additional details.}

\item{seed_radius}{restricts the maximum length of an edge in the graph that
summaries the distance information. \code{NULL} indicates
no restriction. See below for details on how this translates
to maximum distance within a cluster.}

\item{primary_radius}{...}

\item{secondary_radius}{restricts the maximum distance when assigning secondary data
points to clusters.}

\item{batch_size}{...}
}
\value{
Returns a scclust cluster object containing the derived clustering.
}
\description{
\code{sc_clustering} derives a clustering statisfying the specified
size constraints. It implements an algorithm that first summaries the
distance information between data points in a sparse graph and then
constructs the clusterings based on the graph. The function is intended
to run fast while ensuring near-optimal performance.
}
\seealso{
\code{\link{hierarchical_clustering}} can be used to refine the clustering
         constructed by this function.

Other clustering functions: \code{\link{hierarchical_clustering}}
}
\keyword{cluster}
