% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scclust-package.R
\docType{package}
\name{scclust-package}
\alias{scclust-package}
\title{scclust: R wrapper for the scclust library}
\description{
This package wraps the \code{scclust} C library for R. \code{scclust}
provides functions for size constrained clustering. Besides direct wrapper
function, the package provides convenience and utility functions.
}
\details{
The clustering functions are \code{\link{sc_clustering}} and
\code{\link{hierarchical_clustering}}. To get statistics about a clustering
use \code{\link{get_scclust_stats}}. To construct a clustering object from
an existing clustering see \code{\link{scclust}}. To check if a clustering
satisfies some clustering constraints use \code{\link{check_clustering}}.

See the package's website for more information:
\url{https://github.com/fsavje/scclust-R}.

More information about the \code{scclust} C library is found here:
\url{https://github.com/fsavje/scclust}.

Bug reports and suggestions are greatly appreciated. They
are best reported here:
\url{https://github.com/fsavje/scclust-R/issues}.
}
